+++ SpringBackend/src/main/java/org/falcon/TicTacToeCMD/Player.java
@@ -1,4 +1,4 @@
-package org.falcon.TicTacToeLogic;
+package org.falcon.TicTacToeCMD;
 public class Player {
     char playerPiece;
     String name;
