+++ SpringBackend/src/main/resources/static/notes/ANNOTATIONS.md
@@ -1,0 +1,5 @@
+@Component - Talking in reference to a class where the annotation is placed 
+             above:
+A box of object Spring specifies
+https://www.youtube.com/watch?v=K43qyHJXmWI
+WATCH THIS LATER -- SEEMS GOOD
