+++ SpringBackend/src/main/resources/static/notes/SCHEMA.md
