+++ SpringBackend/src/main/resources/static/notes/TROUBLESHOOT.md
@@ -1,0 +1,2 @@
+-----------------Passed, true: ✔    |    Failed, false: ✘--------------------
+*
